Bill Book Changes

+++ src/main/java/com/juri/printing/dao/Prnt01PageDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.juri.printing.entity.Prnt01Page;
 
 @Repository
-public interface Prnt01PageDao extends JpaRepository<Prnt01Page, Integer>{
+public interface Prnt01PageDao extends JpaRepository<Prnt01Page, Long>{
 
 	@Query("select e from Prnt01Page e where e.prnt01PageSize = :paperSize and e.prnt01PageGsm = :gsm")
 	Prnt01Page findByPrnt01PageSizeAndPrnt01PageGsm(@Param("paperSize") String paperSize, @Param("gsm") String gsm);
